+++ src/components/Cart/OrderSummary.js
@@ -1,0 +1,39 @@
+import { useDispatch, useSelector } from "react-redux";
+import { cartActions } from '../../store/cart-slice';
+
+const OrderSummary = () => {
+    const cart = useSelector(state => state.cart);
+
+    const dispatch = useDispatch();
+
+      const cartTotalPrice = cart.itemsList.reduce((accumulator, item) => {
+        return accumulator + item.totalPrice;
+      }, 0);
+    
+      const handleCheckoutClick = () => {
+        dispatch(cartActions.setShowCheckout());
+      };
+    
+  return (
+    <div className='lg:w-1/4 w-full bg-gray-100 p-6 rounded-lg shadow'>
+      <h3 className='font-bold mt-3 mb-5 text-lg'>Summary</h3>
+      <div className='mb-4'>
+        Subtotal <span className='ml-10'>{cartTotalPrice} $</span>
+      </div>
+      <div className='mb-4'>
+        Shipping <span className='font-bold ml-10'> FREE</span>
+      </div>
+      <hr className='my-4 border-t-2 border-gray-300' />
+      <div className='py-2 text-lg font-semibold mb-4'>
+        Total price: {cartTotalPrice} $
+      </div>
+      <button
+        className='btn hover:bg-accent-800 hover:text-white text-black rounded-full w-full py-3'
+        onClick={handleCheckoutClick}>
+        Proceed to checkout
+      </button>
+    </div>
+  );
+};
+
+export default OrderSummary;
